fix: TileInfo 도메인 TestCode 구현완료 및 동작 가능하게 다른 도메인 TestCode 수정

+++ src/main/java/com/mtvs/devlinkbackend/channel/entity/Position.java
@@ -8,7 +8,7 @@ public class Position {
     private float y;
     private float z;
 
-    public Position(int x, int y, int z) {
+    public Position(float x, float y, float z) {
         this.x = x;
         this.y = y;
         this.z = z;
